remove tags.rs
